Introduced module structure/OO principles.
