+++ background.js
@@ -1,35 +1,45 @@
-// This section creates the context menu item and corresponding listener
-async function getCurrentTabUrl() {
+import {getCurrentTabUrl} from './modules/tabs.js';
+import {apiGetPageTickets} from './modules/api_calls.js';
+
+async function getCurrentTab() {
   let queryOptions = { active: true, currentWindow: true };
   let [tab] = await chrome.tabs.query(queryOptions);
-  return tab.url;
+  return tab;
 }
 
-//console.log(await getCurrentTabUrl());
+function getAmountTicketsOnPage() {
+  let queryOptions = { active: true, currentWindow: true };
+  chrome.tabs.query(queryOptions, function(tabs){
 
-let contextMenuItem = {
-  "id": "reportBugItem",
-  "title": "Report bug",
-  "contexts": ["page", "selection"]
-};
+  });
+}
 
-chrome.contextMenus.removeAll(function() {
-  chrome.contextMenus.create(contextMenuItem);
-});
+async function main(){
+  let contextMenuItem = {
+    "id": "reportBugItem",
+    "title": "Report bug",
+    "contexts": ["page", "selection"]
+  };
 
-chrome.contextMenus.onClicked.addListener(function(clickData){
-  //console.log(await getCurrentTabUrl);
-  if (clickData.menuItemId == "reportBugItem"){
-    chrome.tabs.query({currentWindow: true, active: true}, function(tabs){
-      chrome.storage.local.set({ "bug_url": tabs[0].url })
-    });
-    chrome.windows.create({'url': 'reportbug.html', 'type': 'popup'
-    , "height": 720, "width": 600}, function(window) {})
-  }
-});
+  chrome.contextMenus.removeAll(function() {
+    chrome.contextMenus.create(contextMenuItem);
+  });
+  
+  chrome.contextMenus.onClicked.addListener(function(clickData){
+    if (clickData.menuItemId == "reportBugItem"){
+      chrome.tabs.query({currentWindow: true, active: true}, function(tabs){
+        chrome.storage.local.set({ "bug_url": tabs[0].url })
+      });
+      chrome.windows.create({'url': 'reportbug.html', 'type': 'popup'
+      , "height": 720, "width": 600}, function(window) {})
+    }
+  });
+}
+
+main();
 
 // This section creates the badge (number of bugs on page)
-/* function getBadgeStatus() {
+function getBadgeStatus() {
   const status = chrome.storage.sync.get("badgeEnabled", (data) => {
     return data.badgeEnabled
   });
@@ -73,4 +83,4 @@ chrome.storage.onChanged.addListener(function(){
 
 chrome.tabs.onActivated.addListener(function(){
   setBadge();
-}); */
+});
