+++ modules/tabs.js
@@ -1,0 +1,7 @@
+// Module for tab related functions
+
+export async function getCurrentTabUrl() {
+    let queryOptions = { active: true, currentWindow: true };
+    let [tab] = await chrome.tabs.query(queryOptions);
+    return tab.url;
+}
