+++ overview.html
@@ -48,6 +48,6 @@
         </div>
     </div>
     
-    <script src="overview.js"></script>
+    <script type="module" src="overview.js"></script>
 </body>
 </html>
