+++ overview.js
@@ -1,35 +1,12 @@
-async function getJSON(url) {
-    try {
-        const response = await fetch(url, {method: 'GET'});
-  
-        return response.json();
-    } catch (error) {
-        console.error(error);
-    }
-}
-
-async function getCurrentTabUrl() {
-    let queryOptions = { active: true, currentWindow: true };
-    let [tab] = await chrome.tabs.query(queryOptions);
-    return tab.url;
-}
+import {apiGetPageTickets} from './modules/api_calls.js';
 
 async function main() {
 
     let table = document.getElementById("overviewTable").getElementsByTagName('tbody')[0];
-    let currentTabUrl = await getCurrentTabUrl();
-    // URL is encoded twice
-    currentTabUrl = encodeURIComponent(currentTabUrl);
-    currentTabUrl = encodeURIComponent(currentTabUrl);
 
-    const configUrl = chrome.runtime.getURL('/config.json');
-    let apiTicketUrl = await getJSON(configUrl);
-    apiTicketUrl = apiTicketUrl['url_api_ticket_page'];
-    apiTicketUrl = apiTicketUrl + currentTabUrl;
-    
-    let tickets = await getJSON(apiTicketUrl);
-    tickets = tickets['data'];
+    const tickets = await apiGetPageTickets();
 
+    // Fill table with fetched tickets
     for (const ticket in tickets) {
         // Create an empty <tr> element and add it to the 1st position of the table:
         let row = table.insertRow(-1);
