+++ reportbug.html
@@ -63,6 +63,6 @@
 
     </div>
 
-    <script src="reportbug.js"></script>
+    <script type="module" src="reportbug.js"></script>
   </body>
 </html>
