+++ reportbug.js
@@ -1,23 +1,22 @@
 // reportbug.js
+import {apiGetAsJSON} from "./modules/api_calls.js";
 
-// Get URL of the current active user window/tab
-chrome.storage.local.get("bug_url", (data) => {
-    let bug_url = data.bug_url;
-    // URL is encoded twice
-    bug_url = encodeURIComponent(bug_url);
-    bug_url = encodeURIComponent(bug_url);
-    document.getElementById("url-input").value = bug_url;
-});
+async function main() {
 
-// Set form action destination URL
-function setUrl() {
-    if (xhr.readyState === 4) {
-        document.getElementById('report-form').action = 
-            JSON.parse(xhr.response)['url_api_ticket'];
-    }
+    // Get form api url from config file
+    const configUrl = chrome.runtime.getURL('/config.json');
+    let apiTicketUrl = await apiGetAsJSON(configUrl);
+    apiTicketUrl = apiTicketUrl['url_api_ticket'];
+    document.getElementById('report-form').action = apiTicketUrl;
+
+    // Get URL of the current active user window/tab
+    chrome.storage.local.get("bug_url", (data) => {
+        let bug_url = data.bug_url;
+        // URL is encoded twice
+        bug_url = encodeURIComponent(bug_url);
+        bug_url = encodeURIComponent(bug_url);
+        document.getElementById("url-input").value = bug_url;
+    });    
 }
 
-let xhr = new XMLHttpRequest();
-xhr.onreadystatechange = setUrl;
-xhr.open("GET", chrome.runtime.getURL('/config.json'), true);
-xhr.send();
+ main();
